Deletes voice policy dir and shifts it into voiceover dir
