+++ core/feconf.py
@@ -167,7 +167,6 @@ class ValidModelNames(enum.Enum):
     TOPIC = 'topic'
     TRANSLATION = 'translation'
     USER = 'user'
-    VOICE_POLICY = 'voice_policy'
     VOICEOVER = 'voiceover'
