+++ core/platform/models.py
@@ -191,10 +191,6 @@ class _Gae(Platform):
             elif name == Names.USER:
                 from core.storage.user import gae_models as user_models
                 returned_models.append(user_models)
-            elif name == Names.VOICE_POLICY:
-                from core.storage.voice_policy import (
-                    gae_models as voice_policy_models)
-                returned_models.append(voice_policy_models)
             elif name == Names.VOICEOVER:
                 from core.storage.voiceover import (
                     gae_models as voiceover_models)
