+++ mypy_imports.py
@@ -71,7 +71,6 @@ from core.storage.suggestion import gae_models as suggestion_models
 from core.storage.topic import gae_models as topic_models
 from core.storage.translation import gae_models as translation_models
 from core.storage.user import gae_models as user_models
-from core.storage.voice_policy import gae_models as voice_policy_models
 from core.storage.voiceover import gae_models as voiceover_models
 
 __all__ = [
@@ -116,6 +115,5 @@ __all__ = [
     'translate_services',
     'transaction_services',
     'user_models',
-    'voice_policy_models',
     'voiceover_models'
 ]
