+++ scripts/backend_test_shards.json
@@ -395,7 +395,6 @@
         "core.jobs.batch_jobs.story_node_jobs_test",
         "core.storage.voiceover.gae_models_test",
         "core.domain.voiceover_domain_test",
-        "core.domain.voiceover_services_test",
-        "core.storage.voice_policy.gae_models_test"
+        "core.domain.voiceover_services_test"
     ]
 }
